BUG: Update MenuTitle whenever Title is changed and the value of Title used to be.  Bubble the change to update LHS tree. (Trac #7507)
We had to use .data() to track OrigVal manually, but this means that the MenuTitle can be updated much more reliably.  In addition, we trigger the change-handler for MenuTitle that updates the LHS tree.  Unfortunately, we couldn't simply use .trigger('change') because it broken in IE8 due to an entwine bug.