Stop MenuTitle updating tree on change
If the MenuTitle is changed and the tree updates, not saving the
changes to the page leaves the tree with the new title instead of the
old, correct one (until a refresh). This also seems counter-intuitive
anyway - suggesting to users that their updated MenuTitle has been
saved before it actually has.