BUG: Fixed pagination functionality on root assets folder
Moving to another page in the GridField for the root assets folder in AssetAdmin
caused all files to be displayed. This was caused by the functionality that tries
to fix some edge cases for detail view kicked in and didn't filter the list properly.

Fixed by making the edge case detection a little more specific (i.e. checking if
a numeric ID was sent as a request variable).