BUG Page urlsegment field not initialized properly after editing grid field item
No API changes. Formerly when editing a grid field item and then returning to the Main Content tab, the URLSegment field was not properly initialized due to element visibility issues. This makes minor adjustments to the setup of the field to ensure it is always setup in the proper state.