Update code/model/ErrorPage.php
removed some erroneous @params in comments and added onBeforeWrite, which increases the sort number of any ErrorPage to one million.  We do this so that ErrorPages will always show at the bottom of the sitetree (and not in amongst other pages). We could look at them being either at the top or the bottom, but bottom seems most appropriate.