FIX Use SSViewer.theme_enabled in CMS logic
The previously used SSViewer::current_custom_theme() has been
removed since we can't influence the setting of configuration values.

The 'theme_enabled' toggle is a cleaner solution to the same problem,
since the 'custom_theme' was really just a way to remember the original
theme, while still disabling it. The toggle makes this more explicit,
but also requires users of the 'theme' setting to check for it.