FIX: Put self.version in composer dependencies.
Now that https://github.com/composer/composer/pull/1883 is in Composer, self.version
will work as a requirement for framework & cms.  This will simplify the release
process a great deal.

Ultimately, the release of rc1 will be the place to test that, but it seems appropriate
to get this in there for that.

If it succeeds with 3.1-rc1, I'd suggest we backport to 3.0 and 2.4.