NEW: Added post-install and post-update dev/build call.
With this change, composer will automatically call dev/build after an update or install
command.  The main reason that this is 3.1-only is that previous releases would have
necessitated the use of 'sudo -u www-data', which is too environment-specific.

One thing that this may cause problems with is 'composer create-project' in an environment
without a _ss_environment file.  It will call dev/build, and dev/build will throw an
error.  That should be fixed in a subsequent commit.