Behat: Removed unnecessary access-user-help.feature
It doesn't add much value in terms of testing
(can't see how we'd ever mess up the "Help" menu entry),
and it relies on the availability of an external resource
(userhelp.ss.org), making the tests more brittle.

It also breaks the feature following this one, because
somehow Behat doesn't reset the browser session, so
subsequent logins fail because a login still exists.