FIX: Fix failover to index template in ContentController::getViewer()
Historically, if you visit a foo action on Page, and Page_foo.ss doesn't exist, then it fails over
to Page.ss.  The introduction of ContentController::getViewer() broke this, but this patch adds a
test for this case and fixes it.

It was identified by build failures on silverstripe/userforms when tested against the master branch.