NEW: Move from tabs to spaces, initial PSR-2 linting.
This commit adds the “composer run-script lint” command which validates
against a subset of the PSR-2 rules. These are the rules that can be
corrected with phpcbf as a starting point.

It’s expected that the PSR-2 exclusions will be refactored one by one
until we’re fully compliant. In the meantime, this linting stops the
already fixed issues from rotting.