MINOR: concept for a "add CMS Fields" shorthand
It usually takes a fair amount of time to add fields to the CMS for a page as they are not scaffolded. Doing so via shorthand method as suggested will speed this up a fair bit.  What is more, by using the scaffolding, you are less likely to make mistakes, etc...