Enforce canCreate() in "add page" controller
It's already enforced when processing the form,
so there's no reason to list those page types in the first place.
This is a common way to make the CMS less cluttered,
e.g. by creating a DataExtension to remove VirtualPage
from the page types unless you're an admin.

Technically, this changes the CMSMain->PageTypes() API,
but in practice it's only used in CMSPageAddController.
It already respects need_permission, which is described as follows:
> List of permission codes a user can have to allow a user to create a page of this type.

I'm considering the lack of canCreate() checks a bug in this context.