FIX pull up anchors that have been added for linking to
When editing a page an author may insert anchors into their content (via
TinyMCE), and wish to link to those anchors from another point in that
page. E.g. creating a table of contents, a "jump to top" link at the
base of the content, etc.

Previously this required the page to first be saved as a draft, then
reloaded before the anchors would show in the list to link to. Now they
are dynamically pulled from the content entered for a more fluid editing
experience.