NEW Add preview state for non-versioned DataObjects.
Currently only Versioned DataObjects are enabled for the CMS preview.
Without a navigator item, non-versioned DataObjects aren't allowed to
display in the preview panel even if they implement CMSPreviewable.

Note that this state will still need to be added to the list of allowed
states in the .cms-preview entwine logic which can be done like so:
`$('.cms-preview').entwine('ss.preview').AllowedStates().push('Unversioned');`