FIX SiteTreeLinkTracking_Parser should use the getAnchorsOnPage
Currently, the SiteTreeLinkTracking_Parser only checks the page's content for anchors.
As a result any anchors that have been added or modified by the updateAnchorsOnPage extension in the getAnchorsOnPage method are marked ss-broken.

This change updates SiteTreeLinkTracking_Parser to get the anchors from the getAnchorsOnPage function on SiteTree.
This will allow for more consistent RegEx matching and allow the updateAnchorsOnPage extension to be used when checking for broken links.